+++ home-manager-modules/alacritty/default.nix
@@ -62,7 +62,7 @@ in
         unfocused_hollow = true;
       };
 
-      live_config_reload = true;
+      general.live_config_reload = true;
     };
   };
 }