+++ home-manager-modules/nextcloud-client/default.nix
@@ -1,5 +1,6 @@
-_: {
-  services.nextcloud-client = {
+{ pkgs, lib, ... }:
+{
+  services.nextcloud-client = lib.mkIf pkgs.stdenv.isLinux {
     enable = true;
     startInBackground = true;
   };